Add files via upload
This is a getopt c function example.  The arguments are pass to the program as follow.  ./minicalc -a 12 -s 1 -d 1 -m 5 -x 44.   -a adds -s subtracts, -d divides, -m multiplies, -x squares  the last number is the value on which this operations will be performed in the mathematical order starting with the square.